Refactor TNS folder and file, delete useless code and handler.
