+++ skyportal/handlers/api/__init__.py
@@ -200,16 +200,13 @@ from .sysinfo import SysInfoHandler
 from .taxonomy import TaxonomyHandler
 from .telescope import TelescopeHandler
 from .thumbnail import ThumbnailHandler, ThumbnailPathHandler
-from .tns import (
-    BulkTNSHandler,
-    ObjTNSHandler,
-    SpectrumTNSHandler,
-    TNSRobotCoauthorHandler,
-    TNSRobotGroupAutoreporterHandler,
-    TNSRobotGroupHandler,
-    TNSRobotHandler,
-    TNSRobotSubmissionHandler,
-)
+from .tns.obj_tns import ObjTNSHandler
+from .tns.spectrum_tns import SpectrumTNSHandler
+from .tns.tns_robot import TNSRobotHandler
+from .tns.tns_robot_coauthor import TNSRobotCoauthorHandler
+from .tns.tns_robot_group import TNSRobotGroupHandler
+from .tns.tns_robot_group_autoreporter import TNSRobotGroupAutoreporterHandler
+from .tns.tns_robot_submission import TNSRobotSubmissionHandler
 from .unsourced_finder import UnsourcedFinderHandler
 from .user import (
     UserHandler,
